WIP: auto complie soy templates when they change

+++ src/tic/web/closure/__init__.py
@@ -1,10 +1,11 @@
-import depstree
-import jscompiler
-import source
-import treescan
-import closurebuilder
 import logging
+
+import closurebuilder
+import depstree
 from tic import loader
+from tic.core import Component
+from tic.core import implements
+from tic.tools.api import IDirectoryWatcher
 
 def calculate_test_deps(js_test_file_path):
     """
@@ -37,7 +38,7 @@ def calculate_deps(js_entrypoint_file_path):
 
     return [loader.get_relative_path(js_source.GetPath()) for js_source in deps]
 
-def compile_soy_templates(templates=None):
+def compile_soy_templates(templates=None, delete_generated=True):
     """
     Compiles the soy files
     TODO: cannot do this in appengine dev server since os.popen is not defined
@@ -66,11 +67,12 @@ def compile_soy_templates(templates=None):
     logging.info('Found %s template(s)' % len(template_files))
 
     try:
-        shutil.rmtree(generated_path)
+        if delete_generated:
+            shutil.rmtree(generated_path)
+        os.makedirs(generated_path)
     except OSError:
-        pass #no such file or dir
+        pass #no such file or dir, dir already exists 
     
-    os.makedirs(generated_path)
 
     logging.info('compiling ...')
     a = os.popen("java -jar %(soy_to_js_compiler)s --outputPathFormat %(generated_path)s{INPUT_FILE_NAME_NO_EXT}.js %(options)s %(templates)s"
@@ -92,7 +94,49 @@ def compile_soy_templates(templates=None):
     logging.info('Copying required files...')
     jsutil_file = '%s/../tools/closure-templates/soyutils_usegoog.js' % loader.root_path()
     logging.info('\t' + jsutil_file)
-    shutil.copy(jsutil_file, generated_path)
+    shutil.copy(jsutil_file, generated_path) 
     logging.info('Done.')
+
+
+class ClosureTemplatesDirectoryWatcher(Component):
+    implements(IDirectoryWatcher)
+
+    # IDirectoryWatcher API
+    def changed(self, list):
+        """
+        @see IDirectoryWatcher.changed
+        """
+        self._compile(list)
+    def created(self, list):
+        """
+        @see IDirectoryWatcher.created
+        """
+        self._compile(list)
+        
+    def deleted(self, list):
+        """
+        @see IDirectoryWatcher.deleted
+        """
+        #TODO: delete corresponding template
+    # private
+    def _pick_soy_files(self, files):
+        """
+        Removes all non soy files
+        """
+        soys = []
+        for file in files:
+            if file.endswith('.soy'):
+                soys.append(file)
+        return soys
+
+    def _compile(self, files):
+        """
+        compiles the template files
+        """
+        compile_soy_templates(self._pick_soy_files(files), False)
+        
+
+
+
